Rename the hero component to home page hero component
