+++ components/HomePageHero.vue
